refactor: 애니메이션 등록 및 시작 타이밍 제어
1. start 함수를 통해 시작 타이밍 제어
2. 등록은 Main.js에서 일괄 진행
