+++ blue/.vite/deps/_metadata.json
@@ -1,0 +1,8 @@
+{
+  "hash": "2e4381e3",
+  "configHash": "3fcb9768",
+  "lockfileHash": "e3b0c442",
+  "browserHash": "25c53bda",
+  "optimized": {},
+  "chunks": {}
+}
