+++ blue/bassScene/.gitignore
