+++ blue/bassScene/public/models/bass-neck.glb
