--- blue/week1/.gitignore
@@ -1,17 +1,0 @@
-node_modules/
-**/node_modules/
-
-dist/
-**/dist/
-
-build/
-**/build/
-
-.DS_Store
-**/.DS_Store
-
-.vscode/
-**/.vscode/
-
-.env
-**/.env
