--- blue/week1/README.md
@@ -1,14 +1,0 @@
-## 💬 three.js 스터디 1주차
-
-### ✅ 학습 목표
-
-- Three.js의 역할과 구조 이해
-- 개발 환경 세팅 (Vite 등)
-- 기본 Scene, Camera, Renderer 구성
-
-### 🧩 과제
-
-- 회색 배경에 파란색 큐브 하나 띄우기
-  (카메라는 큐브를 정면에서 바라보게)
-
-https://github.com/user-attachments/assets/6ea47876-add1-41fc-8067-f6b2b5f0215f
