--- blue/week2/README.md
@@ -1,15 +1,0 @@
-## 💬 three.js 스터디 2주차
-
-### ✅ 학습 목표
-
-- Mesh란 무엇인가
-- Geometry 종류 살펴보기 (Box, Sphere 등)
-- Material 종류 기본 (MeshBasicMaterial, MeshStandardMaterial)
-
-### 🧩 과제
-
-- "다양한 Material을 적용한 여러 개의 도형 만들기"
-  (Box, Sphere, Cone 등 3개 만들고, 각각 다른 Material 사용)
-
-https://github.com/user-attachments/assets/3baac481-6ba9-49b7-93f2-396b4f790235
-
