--- blue/week3/constants/systemConstants.js
@@ -1,27 +1,0 @@
-export const SYSTEM_CONSTANTS = {
-  BODY_DEPTH: 500,
-  MESH_SCALE: 0.01,
-};
-
-export const STRING_CONSTANTS = [
-  {
-    start: { x: 8.8, y: -14.5, z: 8.025 },
-    end: { x: -10.2, y: 38.5, z: 2.025 },
-    scale: 0.15,
-  },
-  {
-    start: { x: 8.8, y: -14.5, z: 6.675 },
-    end: { x: -10.2, y: 38.5, z: 0.675 },
-    scale: 0.12,
-  },
-  {
-    start: { x: 8.8, y: -14.5, z: 5.325 },
-    end: { x: -10.2, y: 38.5, z: -0.675 },
-    scale: 0.09,
-  },
-  {
-    start: { x: 8.8, y: -14.5, z: 3.975 },
-    end: { x: -10.2, y: 38.5, z: -2.025 },
-    scale: 0.06,
-  },
-];
