--- blue/week3/package.json
@@ -1,8 +1,0 @@
-{
-  "dependencies": {
-    "three": "^0.176.0"
-  },
-  "devDependencies": {
-    "vite": "^6.3.5"
-  }
-}
